7 - Forbid deleting Users with Articles

+++ src/Service/User/Deleter.php
@@ -3,6 +3,7 @@
 namespace App\Service\User;
 
 use App\Repository\UserRepository;
+use App\Service\User\Exception\UserHasArticlesException;
 use App\Service\User\Exception\UserNotFoundException;
 use Doctrine\ORM\EntityManagerInterface;
 
@@ -17,6 +18,10 @@ class Deleter
     {
         if (!($user = $this->userRepository->find($id))) {
             throw new UserNotFoundException('User not found');
+        }
+
+        if ($user->getArticles()->count() > 0) {
+            throw new UserHasArticlesException('Cannot delete user with articles');
         }
 
         $this->entityManager->remove($user);
